变更变量类型

+++ src/FishValueExtension.cs
@@ -343,16 +343,16 @@ namespace System
 		/// </summary>
 		/// <param name="value">要压缩的 <see cref="T:System.Int32"/></param>
 		/// <returns>压缩后的字符串</returns>
-		public static string CompressToString(this int value)
+		public static string CompressToString(this uint value)
 		{
 			var map = new List<char>();
-			var step = _numberCode.Length;
+			var step = (uint)_numberCode.Length;
 
 			while (value > 0)
 			{
 				var digit = value % step;
 				value = (value - digit) / step;
-				map.Add(_numberCode[digit]);
+				map.Add(_numberCode[(int)digit]);
 			}
 
 			return new string(map.Reverse<char>().ToArray());
